Major Changes
Added a countdown with sound, added player and ai vehicle sound as well as menu music.
Implemented pause menu and a menu for when the player finishes the track which displays their best lap time.
Improved ai by adding braking and braking sectors on corners so that it is less likely to crash when going around a corner.
Improved the fluidity between levels and menus so that the player can easily go from the game to the menu and vice versa.
Also did aesthetics updates such as adding shadows to text, minimaps, button backgrounds.